backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
GitPython
openai
requests>=2.28.0
urllib3>=1.26.0